add base project structure with defaul locations.
api.py is the main file defining cave api as well a module loading
config.py is the main file containing the configurations and respective loaders
governor.py is the command line application were we store maintaince scripts.
setup.py is the main file to install cave as a package and define the entry point
for the app manager/governor.
requirements.txt is where we keep our application dependencies
a second readme is added in machinery to help document the directory